requirements.txt
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.169.0
Flask==3.0.0
